Add tile-based SAM segmentation and FIX Missing Sam Parameters
Introduces tile-based segmentation in the SAM model for improved handling of large images, refactors SAMModel for generator creation, and updates the SAMStrategy to support tiled analysis. The analysis and upload routes are modified to allow users to enable tile-based analysis via a new checkbox in the UI.
Missing Sam Parameters : The singleton didn't take new Sam Parameters. The parameters can now be updated for each user.