1/2 : Show object link details in map popups
Adds retrieval and display of object data for linked objects on the map. Popups for object links now show cropped images and metadata, improving visualization and context for users. Updates backend, template, and JS to pass and use object data for popups.
Please, this is in beta, there is still some issues ->
Futur patch :
If two objects link to the same image, only one of them is currently displayed, so the information is not complete.

For example, if one image contains object A and another image also contains object A, then both images will correctly be linked to object A. However, the two images will also be linked to each other. This behavior will make sense for the future visualization, but not for the current version.